made most dependencies optional with Requires.jl
