+++ Project.toml
@@ -6,27 +6,14 @@ version = "0.1.0"
 [deps]
 CSV = "336ed68f-0bac-5ca0-87d4-7b16caf5d00b"
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
-JDF = "babc3d20-cd49-4f60-a736-a8f9c08892d3"
-LibPQ = "194296ae-ab2e-5f79-8cd4-7183a0a5a0d1"
-Parquet = "626c502c-15b0-58ad-a749-f091afb673ae"
-SQLite = "0aa819cd-b072-5ff4-a722-6bc24af294d9"
-StatFiles = "1463e38c-9381-5320-bcd4-4134955f093a"
+Requires = "ae029012-a4dd-5104-9daa-d747884805df"
 Tables = "bd369af6-aec1-5ad0-b16a-f7cc5008161c"
-XLSX = "fdbf4ff8-1666-58a4-91e7-1b58723a45e0"
-ZipFile = "a5390f91-8eb1-5f08-bee0-b1d1ffed6cea"
 
 [compat]
 CSV = "0.7"
 DataFrames = "0.21"
-JDF = "0.2"
-LibPQ = "1.5"
-Parquet = "0.6"
-SQLite = "1"
 Tables = "1"
-XLSX = "0.7"
-ZipFile = "0.9"
 julia = "1"
-StatFiles = "0.8"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
