+++ test/Project.toml
@@ -1,7 +1,11 @@
 [deps]
 DataFrames = "a93c6f00-e57d-5684-b7b6-d8193f3e46c0"
 Dates = "ade2ca70-3891-5945-98fb-dc099432e06a"
+JDF = "babc3d20-cd49-4f60-a736-a8f9c08892d3"
 LibPQ = "194296ae-ab2e-5f79-8cd4-7183a0a5a0d1"
 Parquet = "626c502c-15b0-58ad-a749-f091afb673ae"
 SQLite = "0aa819cd-b072-5ff4-a722-6bc24af294d9"
+StatFiles = "1463e38c-9381-5320-bcd4-4134955f093a"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
+XLSX = "fdbf4ff8-1666-58a4-91e7-1b58723a45e0"
+ZipFile = "a5390f91-8eb1-5f08-bee0-b1d1ffed6cea"
